chore: 拆出useMounted、补全注释
