+++ packages/zent/src/utils/hooks/useAsyncSafeDispatch.ts
@@ -1,5 +1,13 @@
-import { useCallback, useRef, useEffect } from 'react';
+import { useCallback } from 'react';
+import { useMounted } from './useMounted';
 
+/**
+ * 这个hook对React返回的状态更新函数（例
+ * 如useState、useReducer的返回值）包装
+ * 了一层，避免在异步回调中使用时导致可能的
+ * 内存泄漏（Async-Safe）
+ * @param dispatch
+ */
 export function useAsyncSafeDispatch<Args extends unknown[]>(
   dispatch: (...args: Args) => void
 ) {
@@ -10,15 +18,4 @@ export function useAsyncSafeDispatch<Args extends unknown[]>(
     },
     [dispatch, mounted]
   );
-}
-
-function useMounted() {
-  const ref = useRef(false);
-  useEffect(() => {
-    ref.current = true;
-    return () => {
-      ref.current = false;
-    };
-  });
-  return ref;
 }
